Fix profile save requiring two clicks.
After picking a character and clicking "Save profile" the initial click
seems to get eaten, requiring the user to have to click twice.
Only the second click actually triggers the save event, and the notice.
But for some reason setting onBlur to undefined fixes this issue and the
first issue no longer gets eaten. I have no idea why this is, but I
think this fixes it.